Consolidate AnswerFinder into Answers
